combine catrises together with script
